businessseconds/requirements.txt
asgiref==3.2.3
Django==3.0.2
djangorestframework==3.11.0
holidays==0.9.12
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
